feat : globalStyles.js 추가, theme.js 경로 이동, Notosans 폰트 추가, 폰트 적용 세팅
