+++ src/index.js
@@ -6,9 +6,11 @@ import reportWebVitals from './reportWebVitals';
 
 import { Provider } from 'react-redux';
 import store from './modules/configStore';
+import GlobalStyle from './shared/styles/globalStyles';
 
 ReactDOM.render(
   <Provider store={store}>
+    <GlobalStyle />
     <App />
   </Provider>,
   document.getElementById('root'),
